move md5sum and git to tools module
